Memory Overhaul + Fixed Machine Flags

+++ src/operations/Branch.cpp
@@ -27,28 +27,28 @@ void instr_branch(CpuModel * mdl, int32_t offset)
     mdl->setPc(pc);
 }
 
-bool instr_branch_assert(CpuModel * mdl, int8_t test)
+bool instr_branch_assert(CpuModel * mdl, int32_t test)
 {
     int flags = *mdl->getFlagsPtr();
 
     switch(test)
     {
-        /* EQ */ case 0x0: return flags & ZERO;
-        /* NE */ case 0x1: return ! (flags & ZERO);
-        /* HS */ case 0x2: return 0;
-        /* LO */ case 0x3: return 0;
-        /* MI */ case 0x4: return 0;
-        /* PL */ case 0x5: return 0;
-        /* VS */ case 0x6: return 0;
-        /* VC */ case 0x7: return 0;
-        /* HI */ case 0x8: return 0;
-        /* LS */ case 0x9: return 0;
-        /* GE */ case 0xA: return ! (flags & NEGATIVE) || (flags & ZERO);
-        /* LT */ case 0xB: return (flags & NEGATIVE);
-        /* GT */ case 0xC: return ! (flags & NEGATIVE);
-        /* LE */ case 0xD: return (flags & NEGATIVE) || (flags & ZERO);
-    
+        /* EQ */ case 0x0: return flags & ZERO; break;
+        /* NE */ case 0x1: return ! (flags & ZERO); break;
+        /* HS */ case 0x2: return 0; break;
+        /* LO */ case 0x3: return 0; break;
+        /* MI */ case 0x4: return 0; break;
+        /* PL */ case 0x5: return 0; break;
+        /* VS */ case 0x6: return 0; break;
+        /* VC */ case 0x7: return 0; break;
+        /* HI */ case 0x8: return 0; break;
+        /* LS */ case 0x9: return 0; break;
+        /* GE */ case 0xA: return ! (flags & NEGATIVE) || (flags & ZERO); break;
+        /* LT */ case 0xB: return (flags & NEGATIVE); break;
+        /* GT */ case 0xC: return ! (flags & NEGATIVE); break;
+        /* LE */ case 0xD: return (flags & NEGATIVE) || (flags & ZERO); break;
     }
+
 
     return 0;
 }
